backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
aiohttp
python-dotenv
cryptography
google-generativeai
sqlalchemy
requests
schedule
